removed unused util
most of the code int nordicsemi/dfu/util.py seems to be pre python 3.4
related and only one method used in one place.
